+++ src/public/utils/uArray/check.ts
@@ -1,6 +1,7 @@
 /**
  * @Owners zzh
  * @Title public array check方法
+ * @Details 检查一个输入是否为数组
  */
 export const check = <T>(obj: unknown): obj is T[] => obj instanceof Array;
