+++ src/public/utils/uArray/index.ts
@@ -3,17 +3,16 @@
  * @Title public array工具
  */
 import check from './check';
-import jsonParse from './jsonParse';
-import updateAndGetNew from './updateAndGetNew';
 import deleteAndGetNew from './deleteAndGetNew';
 import insertAndGetNew from './insertAndGetNew';
-
+import jsonParse from './jsonParse';
+import updateAndGetNew from './updateAndGetNew';
 
 export const uArray = {
     check,
     jsonParse,
     updateAndGetNew,
     deleteAndGetNew,
-    insertAndGetNew
-}
+    insertAndGetNew,
+};
 export default uArray;
