+++ src/public/utils/uFunction/index.ts
@@ -1,0 +1,15 @@
+/**
+ * @Owners cmZhou
+ * @Title public 函数工具
+ */
+import check from './check';
+import checkExtendsClass from './checkExtendsClass';
+import noop from './noop';
+
+export const uFunction = {
+    check,
+    checkExtendsClass,
+    noop,
+};
+
+export default uFunction;
