+++ src/public/utils/uFunction/noop.ts
@@ -1,0 +1,8 @@
+/**
+ * @Owners zzh
+ * @Title public noop函数
+ */
+
+export const noop = () => { };
+
+export default noop;
