+++ src/public/utils/uHtml/clear.ts
@@ -1,0 +1,8 @@
+/**
+ * @Owners zzh
+ * @Title public clear函数
+ * @Details 去除<>,</>标签
+ */
+export const clear = (html: string) => html.replace(/<\/?[^>]*>/g, '');
+
+export default clear;
