+++ src/public/utils/uHtml/decode.ts
@@ -1,0 +1,9 @@
+/**
+ * @Owners zzh
+ * @Title public decode方法
+ */
+import { decode as htmlDecode } from 'html-entities';
+
+export const decode = htmlDecode;
+
+export default decode;
