+++ src/public/utils/uHtml/encode.ts
@@ -1,0 +1,9 @@
+/**
+ * @Owners zzh
+ * @Title public encode方法
+ */
+import { encode as htmlEncode } from 'html-entities';
+
+export const encode = htmlEncode;
+
+export default encode;
