+++ src/public/utils/uHtml/index.ts
@@ -1,0 +1,16 @@
+/**
+ * @Owners cmZhou
+ * @Title public html工具
+ */
+
+import clear from './clear';
+import decode from './decode';
+import encode from './encode';
+
+export const uHtml = {
+    decode,
+    encode,
+    clear,
+};
+
+export default uHtml;
