+++ src/public/utils/uHttp/createFormData.ts
@@ -1,0 +1,12 @@
+/**
+ * @Owners zzh
+ * @Title public createFormData
+ * @Details 根据对象创建并返回formData格式内容
+ */
+export const createFormData = (obj: dp.Obj<Blob | string>) => {
+    const formData = new FormData();
+    Object.keys(obj).forEach(item => formData.append(item, obj[item]));
+    return formData;
+};
+
+export default createFormData;
