+++ src/public/utils/uNumber/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @Owners zzh
+ * @Title public number工具
+ */
+import check from './check';
+import toFixed from './toFixed';
+
+export const uNumber = {
+    check,
+    toFixed,
+};
+export default uNumber;
