+++ src/public/utils/uObject/deleteKey.ts
@@ -1,0 +1,12 @@
+/**
+ * @Owners zzh
+ * @Title public deleteKey
+ * @Details 删除某个key
+ */
+
+export const deleteKey = <T extends dp.Obj>(obj: T, key: keyof T) => {
+    delete obj[key];
+    return obj;
+};
+
+export default deleteKey;
