+++ src/public/utils/uObject/getSafeJsonObj.ts
@@ -1,0 +1,14 @@
+/**
+ * @Owners zzh
+ * @Title public getSafeJsonObj
+ */
+import parseJson from './parseJson';
+import safeStringify from './safeStringify';
+
+export const getSafeJsonObj = <T extends dp.Obj>(obj: T) => {
+    const res = parseJson<T>(safeStringify(obj));
+    if (res) return res;
+    throw new Error('getSafeJsonObj 执行异常');
+};
+
+export default getSafeJsonObj;
