+++ src/public/utils/uObject/index.ts
@@ -1,0 +1,24 @@
+/**
+ * @Owners zzh
+ * @Title public object工具
+ */
+
+import check from './check';
+import checkInstance from './checkInstance';
+import deleteKey from './deleteKey';
+import getSafeJsonObj from './getSafeJsonObj';
+import parseJson from './parseJson';
+import removeUndefinedProp from './removeUndefinedProp';
+import safeStringify from './safeStringify';
+
+export const uObject = {
+    check,
+    checkInstance,
+    deleteKey,
+    getSafeJsonObj,
+    parseJson,
+    removeUndefinedProp,
+    safeStringify,
+};
+
+export default uObject;
