+++ src/public/utils/uObject/safeStringify.ts
@@ -1,0 +1,8 @@
+/**
+ * @Owners zzh
+ * @Title public stringify
+ * @Details 避免循环引用抛出异常 https://github.com/moll/json-stringify-safe
+ */
+import stringify from 'json-stringify-safe';
+
+export default stringify;
