+++ src/public/utils/uString/equalIC.ts
@@ -1,0 +1,9 @@
+/**
+ * @Owners zzh
+ * @Title public equalIC
+ * @Details 检查两个字符串是否相等（忽略大小写）
+ */
+
+export const equalIC = (a: string, b: string) => a.toLowerCase() === b.toLowerCase();
+
+export default equalIC;
