+++ src/public/utils/uString/index.ts
@@ -1,0 +1,16 @@
+/**
+ * @Owners cmZhou
+ * @Title public string工具
+ */
+
+import check from './check';
+import equalIC from './equalIC';
+import subText from './subText';
+
+export const uString = {
+    check,
+    equalIC,
+    subText,
+};
+
+export default uString;
