+++ src/public/utils/uTask/index.ts
@@ -1,0 +1,14 @@
+/**
+ * @Owners zzh
+ * @Title task util
+ */
+
+import retry from './retry';
+import sleep from './sleep';
+
+export const uTask = {
+    retry,
+    sleep,
+};
+
+export default uTask;
