+++ src/public/utils/uTree/index.ts
@@ -1,0 +1,16 @@
+/**
+ * @Owners zzh
+ * @Title public tree工具
+ */
+
+import create from './create';
+import flatten from './flatten';
+import getPath from './getPath';
+
+export const uTree = {
+    create,
+    flatten,
+    getPath,
+};
+
+export default uTree;
