+++ src/public/utils/uUrl/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @Owners zzh
+ * @Title public url工具
+ */
+
+import ossImg from './ossImg';
+
+export const uUrl = {
+    ossImg,
+};
+export default uUrl;
