+++ src/public/utils/uUuid/index.ts
@@ -1,0 +1,15 @@
+/**
+ * @Owners zzh
+ * @Title uuid util
+ */
+import get from './get';
+import getByNamespace from './getByNamespace';
+import random from './random';
+
+export const uUuid = {
+    get,
+    getByNamespace,
+    random,
+};
+
+export default uUuid;
